requirements.txt
requests
python-dotenv
pytest
pytest-mock
pytest-asyncio
pydantic
langchain-openai
langchain_core
